[CLI] add timeout flag to repo subcommand
