+++ cli/ctl/agent.go
@@ -165,7 +165,7 @@ func listAgentUpgrade(cmd *cobra.Command, args []string) {
 	url := fmt.Sprintf("http://%s:%d/v1/vtaps/", server.IP, server.Port)
 
 	// 调用采集器API，并输出返回结果
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Println(err)
 		return
@@ -202,7 +202,7 @@ func listAgent(cmd *cobra.Command, args []string, output string) {
 	}
 
 	// 调用采集器API，并输出返回结果
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Println(err)
 		return
@@ -267,7 +267,7 @@ func deleteAgent(cmd *cobra.Command, args []string) {
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/vtaps/?name=%s", server.IP, server.Port, args[0])
 	// curl vtap API，get lcuuid
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -277,7 +277,7 @@ func deleteAgent(cmd *cobra.Command, args []string) {
 		lcuuid := response.Get("DATA").GetIndex(0).Get("LCUUID").MustString()
 		url := fmt.Sprintf("http://%s:%d/v1/vtaps/%s/", server.IP, server.Port, lcuuid)
 		// call vtap delete api
-		_, err := common.CURLPerform("DELETE", url, nil, "")
+		_, err := common.CURLPerform("DELETE", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			return
@@ -308,7 +308,7 @@ func updateAgent(cmd *cobra.Command, args []string, updateFilename string) {
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/vtaps/?name=%s", server.IP, server.Port, vtapName)
 	// call vtap api, get lcuuid
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -328,7 +328,7 @@ func updateAgent(cmd *cobra.Command, args []string, updateFilename string) {
 		updateBody["TAP_MODE"] = common.GetVtapTapModeByName(tapMode.(string))
 
 		updateJson, _ := json.Marshal(updateBody)
-		_, err := common.CURLPerform("PATCH", url, nil, string(updateJson))
+		_, err := common.CURLPerform("PATCH", url, nil, string(updateJson), []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			return
@@ -345,7 +345,7 @@ func updateAgent(cmd *cobra.Command, args []string, updateFilename string) {
 	if vtapGroupID, ok := updateMap["vtap_group_id"]; ok {
 		url := fmt.Sprintf("http://%s:%d/v1/vtap-groups/?short_uuid=%s", server.IP, server.Port, vtapGroupID)
 		// call vtap-group api, get lcuuid
-		response, err := common.CURLPerform("GET", url, nil, "")
+		response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			return
@@ -369,7 +369,7 @@ func updateAgent(cmd *cobra.Command, args []string, updateFilename string) {
 	// call vtap update api
 	updateJson, _ := json.Marshal(updateMap)
 	url = fmt.Sprintf("http://%s:%d/v1/vtaps/%s/", server.IP, server.Port, lcuuid)
-	_, err = common.CURLPerform("PATCH", url, nil, string(updateJson))
+	_, err = common.CURLPerform("PATCH", url, nil, string(updateJson), []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -397,7 +397,7 @@ func upgadeAgent(cmd *cobra.Command, args []string) {
 
 	server := common.GetServerInfo(cmd)
 	serverURL := fmt.Sprintf("http://%s:%d/v1/controllers/", server.IP, server.Port)
-	response, err := common.CURLPerform("GET", serverURL, nil, "")
+	response, err := common.CURLPerform("GET", serverURL, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -420,7 +420,7 @@ func upgadeAgent(cmd *cobra.Command, args []string) {
 	}
 
 	vtapURL := fmt.Sprintf("http://%s:%d/v1/vtaps/?name=%s", server.IP, server.Port, vtapName)
-	response, err = common.CURLPerform("GET", vtapURL, nil, "")
+	response, err = common.CURLPerform("GET", vtapURL, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -456,7 +456,7 @@ func upgadeAgent(cmd *cobra.Command, args []string) {
 	}
 	for host, _ := range hosts {
 		url := fmt.Sprintf(url_format, host, server.Port, vtapLcuuid)
-		response, err := common.CURLPerform("PATCH", url, body, "")
+		response, err := common.CURLPerform("PATCH", url, body, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			fmt.Printf("upgrade agent %s server %s failed, response: %s\n", vtapName, host, response)
@@ -468,9 +468,7 @@ func upgadeAgent(cmd *cobra.Command, args []string) {
 }
 
 func rebalance(cmd *cobra.Command, rebalanceType RebalanceType, typeVal string) error {
-	server := common.GetServerInfo(cmd)
-
-	isBalance, err := ifNeedRebalance(server, typeVal)
+	isBalance, err := ifNeedRebalance(cmd, typeVal)
 	if err != nil {
 		return err
 	}
@@ -482,7 +480,7 @@ func rebalance(cmd *cobra.Command, rebalanceType RebalanceType, typeVal string) 
 		fmt.Println("no balance required")
 		return nil
 	}
-	resp, err := execRebalance(server, typeVal)
+	resp, err := execRebalance(cmd, typeVal)
 	if err != nil {
 		return err
 	}
@@ -493,9 +491,10 @@ func rebalance(cmd *cobra.Command, rebalanceType RebalanceType, typeVal string) 
 	return nil
 }
 
-func ifNeedRebalance(server *common.Server, typeStr string) (bool, error) {
+func ifNeedRebalance(cmd *cobra.Command, typeStr string) (bool, error) {
+	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/rebalance-vtap/?check=false&type=%s", server.IP, server.Port, typeStr)
-	resp, err := common.CURLPerform("POST", url, nil, "")
+	resp, err := common.CURLPerform("POST", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		return false, err
 	}
@@ -505,9 +504,10 @@ func ifNeedRebalance(server *common.Server, typeStr string) (bool, error) {
 	return true, nil
 }
 
-func execRebalance(server *common.Server, typeStr string) (*simplejson.Json, error) {
+func execRebalance(cmd *cobra.Command, typeStr string) (*simplejson.Json, error) {
+	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/rebalance-vtap/?check=true&type=%s", server.IP, server.Port, typeStr)
-	resp, err := common.CURLPerform("POST", url, nil, "")
+	resp, err := common.CURLPerform("POST", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		return nil, err
 	}
