+++ cli/ctl/agent_group_config.go
@@ -98,7 +98,7 @@ func RegisterAgentGroupConfigCommand() *cobra.Command {
 func exampleAgentGroupConfig(cmd *cobra.Command, args []string) {
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/vtap-group-configuration/example/", server.IP, server.Port)
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -121,7 +121,7 @@ func listAgentGroupConfig(cmd *cobra.Command, args []string, output string) {
 			url += "advanced/"
 		}
 
-		response, err := common.CURLPerform("GET", url, nil, "")
+		response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			return
@@ -135,7 +135,7 @@ func listAgentGroupConfig(cmd *cobra.Command, args []string, output string) {
 			}
 		}
 	} else {
-		response, err := common.CURLPerform("GET", url, nil, "")
+		response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			return
@@ -163,7 +163,7 @@ func createAgentGroupConfig(cmd *cobra.Command, args []string, createFilename st
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 	}
-	_, err = common.CURLPerform("POST", url, nil, string(yamlFile))
+	_, err = common.CURLPerform("POST", url, nil, string(yamlFile), []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -193,7 +193,7 @@ func updateAgentGroupConfig(cmd *cobra.Command, args []string, updateFilename st
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/vtap-group-configuration/?vtap_group_id=%s", server.IP, server.Port, vtapGroupID)
 	// call vtap-group api, get lcuuid
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -207,7 +207,7 @@ func updateAgentGroupConfig(cmd *cobra.Command, args []string, updateFilename st
 
 	// call vtap-group config update api
 	url = fmt.Sprintf("http://%s:%d/v1/vtap-group-configuration/advanced/%s/", server.IP, server.Port, lcuuid)
-	_, err = common.CURLPerform("PATCH", url, nil, string(yamlFile))
+	_, err = common.CURLPerform("PATCH", url, nil, string(yamlFile), []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -225,7 +225,7 @@ func deleteAgentGroupConfig(cmd *cobra.Command, args []string) {
 		"http://%s:%d/v1/vtap-group-configuration/filter/?vtap_group_id=%s",
 		server.IP, server.Port, args[0],
 	)
-	_, err := common.CURLPerform("DELETE", url, nil, "")
+	_, err := common.CURLPerform("DELETE", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
