+++ cli/ctl/genesis.go
@@ -100,7 +100,7 @@ func syncInfo(cmd *cobra.Command, resType string) {
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/sync/%s/", server.IP, server.Port, resType)
 
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -152,7 +152,7 @@ func k8sInfo(cmd *cobra.Command, args []string, resType string) {
 	server := common.GetServerInfo(cmd)
 	url := fmt.Sprintf("http://%s:%d/v1/kubernetes-info/%s/", server.IP, server.Port, args[0])
 
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -388,7 +388,7 @@ func agentInfo(cmd *cobra.Command, args []string) {
 	}
 
 	url := fmt.Sprintf("http://%s:%d/v1/agent-stats/%s/", podIP, server.SvcPort, args[0])
-	response, err := common.CURLPerform("GET", url, nil, "")
+	response, err := common.CURLPerform("GET", url, nil, "", []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		return
@@ -417,7 +417,7 @@ func prometheusInfo(cmd *cobra.Command, args []string) {
 	}
 
 	path := fmt.Sprintf("/v1/prometheus-info/%s/", args[0])
-	common.GetURLInfo(cmd, path)
+	common.GetURLInfo(cmd, path, []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 }
 
 func storageInfo(cmd *cobra.Command, args []string) {
@@ -427,5 +427,5 @@ func storageInfo(cmd *cobra.Command, args []string) {
 	}
 
 	path := fmt.Sprintf("/v1/genesis-storage/%s/", args[0])
-	common.GetURLInfo(cmd, path)
+	common.GetURLInfo(cmd, path, []common.HTTPOption{common.WithTimeout(common.GetTimeout(cmd))}...)
 }
